+++ lib/unit_measurements/rails/railtie.rb
@@ -4,12 +4,11 @@
 
 module UnitMeasurements
   module Rails
-    # The +Railtie+ class for integrating +unit_measurements+ with the +Rails+
-    # framework.
+    # The +Railtie+ class integrates +unit_measurements+ with the Rails framework.
     #
     # This Railtie is designed to be automatically loaded by Rails when the
-    # application starts. It can be used to configure and customize the behavior
-    # of +unit_measurements+ in a Rails application.
+    # application starts. It facilitates configuration and customization of the
+    # behavior of +unit_measurements+ within a Rails application.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.1.0