+++ lib/unit_measurements/rails/unit_groups/all.rb
@@ -1,3 +1,8 @@
 # -*- encoding: utf-8 -*-
 # -*- frozen_string_literal: true -*-
 # -*- warn_indent: true -*-
+
+require_relative "length"
+require_relative "weight"
+require_relative "volume"
+require_relative "area"