+++ spec/support/models/cube.rb
@@ -1,0 +1,9 @@
+# -*- encoding: utf-8 -*-
+# -*- frozen_string_literal: true -*-
+# -*- warn_indent: true -*-
+
+class Cube < ActiveRecord::Base
+  measured UnitMeasurements::Length, :length, :width
+
+  measured_length :height
+end